+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 * Alert: Achtergrondkleur wordt op verkeerde niveau gezet ([#2614](https://github.com/dso-toolkit/dso-toolkit/issues/2614))
 * Document Component: Titel knop mist aria-expanded en sluit voor titel tekst ([#2480](https://github.com/dso-toolkit/dso-toolkit/issues/2480))
 * Document Component: Ondersteuning lijstnummers IMRO documenten ([#2500](https://github.com/dso-toolkit/dso-toolkit/issues/2500))
+* Document List: Add aria-expanded to dropdown menu example button ([#2561](https://github.com/dso-toolkit/dso-toolkit/issues/2561))
 
 ### Tasks
 * Packages: Dependencies updates ([#2483](https://github.com/dso-toolkit/dso-toolkit/issues/2483))